fix: grpc search tracing
