+++ src/handler/grpc/request/search/mod.rs
@@ -28,9 +28,8 @@ use proto::cluster_rpc::{
     search_server::Search,
 };
 use tonic::{Request, Response, Status};
-use tracing::Instrument;
 
-use crate::service::{search as SearchService, websocket_events::setup_tracing_with_trace_id};
+use crate::service::search as SearchService;
 
 #[derive(Clone, Debug)]
 #[cfg(feature = "enterprise")]
@@ -128,17 +127,13 @@ impl Default for Searcher {
 
 #[tonic::async_trait]
 impl Search for Searcher {
+    #[tracing::instrument(name = "grpc:search:search", skip_all)]
     async fn search(
         &self,
         req: Request<SearchRequest>,
     ) -> Result<Response<SearchResponse>, Status> {
         let start = std::time::Instant::now();
         let req = req.into_inner();
-        let search_span = setup_tracing_with_trace_id(
-            &req.trace_id,
-            tracing::info_span!("src::handler::grpc::request::search::search"),
-        )
-        .await;
 
         let request = json::from_slice::<search::Request>(&req.request)
             .map_err(|e| Status::internal(format!("failed to parse search request: {e}")))?;
@@ -151,7 +146,6 @@ impl Search for Searcher {
             &request,
             "".to_string(),
         )
-        .instrument(search_span)
         .await;
 
         match ret {
@@ -169,16 +163,12 @@ impl Search for Searcher {
         }
     }
 
+    #[tracing::instrument(name = "grpc:search:search_multi", skip_all)]
     async fn search_multi(
         &self,
         req: Request<SearchRequest>,
     ) -> Result<Response<SearchResponse>, Status> {
         let req = req.into_inner();
-        let search_span = setup_tracing_with_trace_id(
-            &req.trace_id,
-            tracing::info_span!("src::handler::grpc::request::search::search_multi"),
-        )
-        .await;
         let request =
             json::from_slice::<search::MultiStreamRequest>(&req.request).map_err(|e| {
                 Status::internal(format!("failed to parse multi-stream search request: {e}"))
@@ -186,7 +176,6 @@ impl Search for Searcher {
         let stream_type = StreamType::from(req.stream_type.as_str());
         let ret =
             SearchService::search_multi(&req.trace_id, &req.org_id, stream_type, None, &request)
-                .instrument(search_span)
                 .await;
 
         match ret {
