+++ src/service/mod.rs
@@ -15,6 +15,8 @@
 
 use config::{meta::stream::StreamParams, utils::schema::format_stream_name};
 use infra::errors::Result;
+use tracing_opentelemetry::OpenTelemetrySpanExt;
+
 pub mod alerts;
 pub mod circuit_breaker;
 pub mod cluster_info;
@@ -64,4 +66,19 @@ pub async fn get_formatted_stream_name(params: StreamParams) -> Result<String> {
     } else {
         stream_name
     })
+}
+
+/// Setup tracing with a trace ID
+pub async fn setup_tracing_with_trace_id(trace_id: &str, span: tracing::Span) -> tracing::Span {
+    let mut headers: std::collections::HashMap<String, String> = std::collections::HashMap::new();
+    let traceparent = format!(
+        "00-{}-{}-01", /* 01 to indicate that the span is sampled i.e. needs to be
+                        * recorded/exported */
+        trace_id,
+        config::ider::generate_span_id()
+    );
+    headers.insert("traceparent".to_string(), traceparent);
+    let parent_ctx = opentelemetry::global::get_text_map_propagator(|prop| prop.extract(&headers));
+    span.set_parent(parent_ctx);
+    span
 }
